chore(mobile): reorganize and move around files
