+++ apps/mobile/src/features/posts/components/avatar.tsx
